requirements.txt
streamlit
google-generativeai
python-docx
PyPDF2
openai
fpdf2
reportlab
pytest